+++ Sources/TraceLog/Proxies/DirectWriterProxy.swift
@@ -34,6 +34,6 @@ internal class DirectWriterProxy: WriterProxy {
 
     @inline(__always)
     internal func log(_ timestamp: Double, level: LogLevel, tag: String, message: String, runtimeContext: RuntimeContext, staticContext: StaticContext) {
-        self.writer.log(timestamp, level: level, tag: tag, message: message, runtimeContext: runtimeContext, staticContext: staticContext)
+        _ = self.writer.log(timestamp, level: level, tag: tag, message: message, runtimeContext: runtimeContext, staticContext: staticContext)
     }
 }