+++ Tests/TraceLogTests/TraceLogPerformanceTests.swift
@@ -14,7 +14,9 @@ private let testIterations = 1000
 class TraceLogPerformanceTestsSwift: XCTestCase {
 
     struct NullWriter: Writer {
-        func log(_ timestamp: Double, level: LogLevel, tag: String, message: String, runtimeContext: RuntimeContext, staticContext: StaticContext) {}
+        func log(_ timestamp: Double, level: LogLevel, tag: String, message: String, runtimeContext: RuntimeContext, staticContext: StaticContext) -> LogResult {
+            return .success
+        }
     }
 
     func testLogErrorPerformance_NullWriter() {